Added some example Run frames
With my changes for controllable frame length generation, I uploaded some random images for demonstration, and a picture of how the numpy arrays that represent each row are stored in separate directories, as requested.  The full directories for one run for each of our models in zipped form, are still stored locally for good measure in my PC in case it is needed (GitHub wouldn't let me upload them due to large size).